Change type of column peptide_instance.elution_time from DOUBLE to FLOAT

+++ Proline-Databases/src/main/resources/dbscripts/msi/h2/V0_1__init_msidb.sql
@@ -676,7 +676,7 @@ CREATE TABLE peptide_instance (
                 protein_match_count INTEGER NOT NULL,
                 protein_set_count INTEGER NOT NULL,
                 selection_level INTEGER NOT NULL,
-                elution_time DOUBLE,
+                elution_time REAL,
                 serialized_properties LONGVARCHAR,
                 best_peptide_match_id INTEGER NOT NULL,
                 peptide_id INTEGER NOT NULL,
